clip_rag_code/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
Pillow
pinecone
pymupdf
python-dotenv
httpx
aiofiles